move auth routers into a separate file
