+++ app/schemas/dancer.py
@@ -18,13 +18,12 @@ class BaseDancer(BaseModel):
     youtube: Optional[str] = None
     agency: Optional[str] = None
     contact_email: Optional[str] = None
-    
+
     model_config = ConfigDict(from_attributes=True)
 
 
 class Dancer(BaseDancer):
     id: UUID4
-
 
 
 class CreateDancer(BaseDancer):
