+++ app/schemas/training.py
@@ -20,14 +20,13 @@ class TrainingBase(BaseModel):
     flyer: Optional[str] = None
     max_slots: int
     is_active: bool
-    
+
     model_config = ConfigDict(from_attributes=True)
 
 
 class Training(TrainingBase):
     id: UUID4
     created_at: datetime
-
 
 
 class CreateTraining(TrainingBase):
